update example for v2
