+++ composer.json
@@ -13,7 +13,7 @@
   "require-dev": {
     "bamarni/composer-bin-plugin": "^1.8.2",
     "phpstan/extension-installer": "^1.2",
-    "phpstan/phpstan": "^1.10.11",
+    "phpstan/phpstan": "^1.10.13",
     "phpstan/phpstan-deprecation-rules": "^1.1.3",
     "phpstan/phpstan-strict-rules": "^1.5.1",
     "rector/rector": "^0.15.24",
@@ -37,6 +37,12 @@
   "autoload": {
     "psr-4": {
       "Valantic\\ElasticaBridgeBundle\\": "src/"
+    }
+  },
+  "autoload-dev": {
+    "psr-4": {
+      "App\\": "docs/example/src/",
+      "Pimcore\\Model\\DataObject\\": "docs/example/var/classes/DataObject/"
     }
   },
   "extra": {
