--- docs/example/src/AppBundle/Elasticsearch/Document/CategoryDocument.php
@@ -1,27 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace AppBundle\Elasticsearch\Document;
-
-use Pimcore\Model\DataObject\Category;
-use Valantic\ElasticaBridgeBundle\DocumentType\AbstractDocument;
-use Valantic\ElasticaBridgeBundle\DocumentType\DocumentInterface;
-
-class CategoryDocument extends AbstractDocument
-{
-    public function getType(): string
-    {
-        return DocumentInterface::TYPE_OBJECT;
-    }
-
-    public function getSubType(): string
-    {
-        return Category::class;
-    }
-
-    public function treatObjectVariantsAsDocuments(): bool
-    {
-        return false;
-    }
-}
