--- docs/example/src/AppBundle/Elasticsearch/Index/Search/Document/SearchIndexAwareTrait.php
@@ -1,20 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace AppBundle\Elasticsearch\Index\Search\Document;
-
-use AppBundle\Elasticsearch\Index\Search\SearchIndex;
-
-trait SearchIndexAwareTrait
-{
-    /**
-     * @param SearchIndex $index
-     *
-     * @required
-     */
-    public function setIndex(SearchIndex $index): void
-    {
-        $this->index = $index;
-    }
-}
