+++ docs/example/src/Controller/Api/SearchController.php
@@ -2,9 +2,10 @@
 
 declare(strict_types=1);
 
-namespace AppBundle\Controller\Api;
+namespace App\Controller\Api;
 
-use AppBundle\Elasticsearch\Index\Product\ProductIndex;
+use App\Elasticsearch\Index\Product\ProductIndex;
+use Elastica\Query;
 use Pimcore\Localization\LocaleService;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\JsonResponse;
@@ -12,12 +13,16 @@ use Symfony\Component\HttpFoundation\Request;
 
 class SearchController extends AbstractController
 {
-    public function productAction(Request $request, ProductIndex $productIndex, LocaleService $localeService): JsonResponse
-    {
+    public function productAction(
+        Request $request,
+        ProductIndex $productIndex,
+        LocaleService $localeService,
+    ): JsonResponse {
+        /** @var string $query */
         $query = $request->query->get('query');
 
         $results = $productIndex->getElasticaIndex()
-            ->search($productIndex->filterByLocaleAndQuery($localeService->getLocale(), $query))
+            ->search(new Query($productIndex->filterByLocaleAndQuery($localeService->getLocale() ?? 'en', $query)))
             ->getDocuments();
 
         return new JsonResponse($results);
