+++ docs/example/src/Elasticsearch/Index/Product/ProductIndex.php
@@ -2,35 +2,24 @@
 
 declare(strict_types=1);
 
-namespace AppBundle\Elasticsearch\Index\Product;
+namespace App\Elasticsearch\Index\Product;
 
-use AppBundle\Elasticsearch\Index\Category\CategoryIndex;
-use AppBundle\Elasticsearch\Index\Product\Document\ProductIndexDocument;
+use App\Elasticsearch\Index\Product\Document\ProductIndexDocument;
 use Elastica\Query\BoolQuery;
 use Elastica\Query\MatchQuery;
 use Elastica\Query\MultiMatch;
-use Pimcore\Model\DataObject\Category;
 use Pimcore\Model\DataObject\Product;
-use Valantic\ElasticaBridgeBundle\DocumentType\Index\DocumentInterface;
-use Valantic\ElasticaBridgeBundle\Elastica\Client\ElasticsearchClient;
+use Valantic\ElasticaBridgeBundle\Document\DocumentInterface;
+use Valantic\ElasticaBridgeBundle\Enum\DocumentType;
 use Valantic\ElasticaBridgeBundle\Index\AbstractIndex;
 use Valantic\ElasticaBridgeBundle\Index\TenantAwareInterface;
 use Valantic\ElasticaBridgeBundle\Index\TenantAwareTrait;
-use Valantic\ElasticaBridgeBundle\Repository\DocumentRepository;
 
 class ProductIndex extends AbstractIndex implements TenantAwareInterface
 {
     use TenantAwareTrait;
 
     public const ATTRIBUTE_CATEGORIES = 'categories';
-    protected CategoryIndex $categoryIndex;
-
-    public function __construct(ElasticsearchClient $client, DocumentRepository $indexDocumentRepository, CategoryIndex $categoryIndex)
-    {
-        parent::__construct($client, $indexDocumentRepository);
-
-        $this->categoryIndex = $categoryIndex;
-    }
 
     public function getTenantUnawareName(): string
     {
@@ -39,13 +28,9 @@ class ProductIndex extends AbstractIndex implements TenantAwareInterface
 
     public function getAllowedDocuments(): array
     {
-        return [ProductIndexDocument::class];
-    }
-
-    public function filterByCategory(Category $category): BoolQuery
-    {
-        return (new BoolQuery())
-            ->addMust(new MatchQuery(self::ATTRIBUTE_CATEGORIES, $category->getId()));
+        return [
+            ProductIndexDocument::class,
+        ];
     }
 
     public function filterByLocaleAndQuery(string $locale, string $query): BoolQuery
@@ -58,7 +43,7 @@ class ProductIndex extends AbstractIndex implements TenantAwareInterface
                     ])
                     ->setQuery($query)
             )
-            ->addFilter(new MatchQuery(DocumentInterface::META_TYPE, DocumentInterface::TYPE_OBJECT))
+            ->addFilter(new MatchQuery(DocumentInterface::META_TYPE, DocumentType::DATA_OBJECT->value))
             ->addFilter(new MatchQuery(DocumentInterface::META_SUB_TYPE, Product::class));
     }
