+++ docs/example/var/classes/DataObject/Category.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Pimcore\Model\DataObject;
+
+class Category extends Concrete
+{
+    public function getId(): int
+    {
+        return 0;
+    }
+}
