+++ docs/example/var/classes/DataObject/Product.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Pimcore\Model\DataObject;
+
+class Product extends Concrete
+{
+}
